requirements.txt
APScheduler
PyEmail
python-dotenv